add Mesh

+++ ppm.h
@@ -2,14 +2,15 @@
 #define __PULSAR_PPM_H__
 
 #include <cstdio>
+#include <cmath>
 
 namespace pulsar{
 	
-	inline double clamp(double x){
+	inline float clamp(float x){
 		return x<0 ? 0 : x>1 ? 1 : x;
 	}
-	inline int toInt(double x){
-		return int(pow(clamp(x), 1/2.2) * 255 + 0.5);
+	inline int toInt(float x){
+		return int(pow(clamp(x), 1.0f/2.2f) * 255 + 0.5f);
 	}
 
 	void save_ppm_file(const char* szFilename, const float* image, const int width, const int height) {
